+++ modules/webgl/src/constants/webgl-constants.ts
@@ -1045,7 +1045,11 @@ enum GLEnum {
   MAX_DUAL_SOURCE_DRAW_BUFFERS_WEBGL = 0x88fc,
 
   /** EXT_texture_mirror_clamp_to_edge https://registry.khronos.org/webgl/extensions/EXT_texture_mirror_clamp_to_edge/ */
-  MIRROR_CLAMP_TO_EDGE_EXT = 0x8743
+  MIRROR_CLAMP_TO_EDGE_EXT = 0x8743,
+
+  /** WEBGL_stencil_texturing https://registry.khronos.org/webgl/extensions/WEBGL_stencil_texturing/ */
+  DEPTH_STENCIL_TEXTURE_MODE_WEBGL = 0x90ea,
+  STENCIL_INDEX_WEBGL = 0x1901
 }
 
 export {GLEnum as GL};